StoryLine
Timer, Hp
When some time left on the timer, the player's health will start decrease.
i deleted from the ray cast script the variables we didnt needed.
