--- Assets/Scripts/AmmoMovement.cs
@@ -1,24 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class AmmoMovement : MonoBehaviour
-{
-    float speed;
-    // Start is called before the first frame update
-    void Start()
-    {
-        speed = 50f;
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        transform.Translate(speed * Time.deltaTime * Vector3.up);
-    }
-
-    private void OnCollisionEnter(Collision collision)
-    {
-        Destroy(gameObject);
-    }
-}
