+++ Assets/Scripts/Health.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f6958ad6058ab6944ba3eb114e6769cc
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
