+++ Assets/Scripts/Health/HealthManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cf742acf91163ab47aed050d0136d839
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
