+++ Assets/Scripts/StoryLine.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1f551c047d413884ca14924a242a8e99
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
