+++ Assets/Scripts/StoryLine/StoryLine.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2c816c43996572d4fbec978d3202a9be
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
