+++ Assets/Scripts/StoryLine/Timer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4a031179b4fc8864fac50d0ab251c9c8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
