+++ Assets/Scripts/Weapons/WeaponBase.cs.meta
